+++ src/lib.rs
@@ -18,6 +18,7 @@ mod path;
 mod pixel;
 mod plotter;
 mod raster;
+mod rgb8;
 mod rgba8;
 mod stroker;
 
@@ -28,5 +29,6 @@ pub use path::{FillRule,Path2D,PathBuilder,PathOp};
 pub use pixel::PixFmt;
 pub use plotter::Plotter;
 pub use raster::{Raster,RasterB};
+pub use rgb8::Rgb8;
 pub use rgba8::Rgba8;
 pub use stroker::JoinStyle;