+++ src/pixel.rs
@@ -10,6 +10,7 @@ use png::ColorType;
 /// as well as the layout of pixels in memory.
 ///
 /// * [Gray8](struct.Gray8.html)
+/// * [Rgb8](struct.Rgb8.html)
 /// * [Rgba8](struct.Rgba8.html)
 pub trait PixFmt: Clone + Default {